+++ ChkdraftLib/Windows/DialogWindows/MapSettings/MapProperties.cpp
@@ -55,7 +55,7 @@ MapPropertiesWindow::~MapPropertiesWindow()
 
 }
 
-bool MapPropertiesWindow::CreateThis(HWND hParent, u32 windowId)
+bool MapPropertiesWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
         return SetParent(hParent);
