+++ ChkdraftLib/Windows/DialogWindows/MapSettings/TechSettings.cpp
@@ -36,7 +36,7 @@ TechSettingsWindow::~TechSettingsWindow()
 
 }
 
-bool TechSettingsWindow::CreateThis(HWND hParent, u32 windowId)
+bool TechSettingsWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
         return SetParent(hParent);
