+++ ChkdraftLib/Windows/DialogWindows/TrigEditor/Switches.cpp
@@ -19,7 +19,7 @@ SwitchesWindow::~SwitchesWindow()
 
 }
 
-bool SwitchesWindow::CreateThis(HWND hParent, u32 windowId)
+bool SwitchesWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
         return SetParent(hParent);
