+++ ChkdraftLib/Windows/DialogWindows/TrigEditor/Triggers.cpp
@@ -42,7 +42,7 @@ TriggersWindow::~TriggersWindow()
 
 }
 
-bool TriggersWindow::CreateThis(HWND hParent, u32 windowId)
+bool TriggersWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
         return SetParent(hParent);
