+++ ChkdraftLib/Windows/DialogWindows/TrigModify/TrigModifyText.cpp
@@ -26,7 +26,7 @@ TrigModifyTextWindow::~TrigModifyTextWindow()
 
 }
 
-bool TrigModifyTextWindow::CreateThis(HWND hParent, u32 windowId)
+bool TrigModifyTextWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
     {
